+++ db/migrate/20200126030829_change_activity_duration_type.rb
@@ -1,0 +1,5 @@
+class ChangeActivityDurationType < ActiveRecord::Migration[6.0]
+  def change
+    change_column :activities, :duration, :float
+  end
+end